fei/requirements.txt
alabaster==0.7.12
amqp==2.6.0
asgiref==3.2.7
astroid==2.3.3
autopep8==1.5
Babel==2.8.0
billiard==3.6.3.0
bitarray==1.2.2
celery==4.4.5
certifi==2019.11.28
chardet==3.0.4
commonmark==0.9.1
Django==2.2.13
django-redis==4.11.0
djangorestframework==3.11.0
docutils==0.16
future==0.18.2
gunicorn==20.0.4
html2text==2020.1.16
idna==2.9
imagesize==1.2.0
importlib-metadata==1.6.1
isort==4.3.21
Jinja2==2.11.1
kombu==4.6.10
lazy-object-proxy==1.4.3
MarkupSafe==1.1.1
mccabe==0.6.1
mypy==0.770
mypy-extensions==0.4.3
mysqlclient==1.4.6
packaging==20.3
Pillow==7.1.2
pycodestyle==2.5.0
pydash==4.7.6
Pygments==2.6.1
pylint==2.4.4
pyparsing==2.4.6
pytz==2019.3
PyYAML==5.3.1
recommonmark==0.6.0
redis==3.5.2
requests==2.23.0
six==1.14.0
snowballstemmer==2.0.0
Sphinx==2.4.4
sphinx-markdown-builder==0.5.4
sphinx-rtd-theme==0.4.3
sphinxcontrib-applehelp==1.0.2
sphinxcontrib-devhelp==1.0.2
sphinxcontrib-htmlhelp==1.0.3
sphinxcontrib-jsmath==1.0.1
sphinxcontrib-qthelp==1.0.3
sphinxcontrib-serializinghtml==1.1.4
sqlparse==0.3.1
typed-ast==1.4.1
typing-extensions==3.7.4.2
unify==0.5
untokenize==0.1.1
uritemplate==3.0.1
urllib3==1.25.8
uWSGI==2.0.18
vine==1.3.0
wrapt==1.11.2
yapf==0.29.0
zipp==3.1.0
